requirements.txt
Django
djangorestframework
Pillow
python-decouple